requirements.txt
streamlit
PyPDF2
pandas
openai
faiss-cpu
python-dotenv
